+++ android/src/main/java/com/pspdfkit/flutter/pspdfkit/EventDispatcher.java
@@ -52,7 +52,9 @@ public class EventDispatcher {
         sendEvent("flutterPdfActivityOnPause");
     }
 
-
+    public void notifyPdfFragmentAdded() {
+        sendEvent("flutterPdfFragmentAdded");
+    }
 
     public void notifyInstantSyncStarted(String documentId) {
         sendEvent("pspdfkitInstantSyncStarted", documentId);