+++ android/src/main/java/com/pspdfkit/flutter/pspdfkit/PSPDFKitView.kt
@@ -75,6 +75,7 @@ internal class PSPDFKitView(
                     .build()
             }
         }
+        getFragmentActivity(context).supportFragmentManager.registerFragmentLifecycleCallbacks(FlutterPdfUiFragmentCallbacks(), true)
 
         fragmentContainerView?.let {
             it.addOnAttachStateChangeListener(object : View.OnAttachStateChangeListener {